feat(LottoRank): getLottoRank 메서드 추가
LottoRank
- matchCount, bonusMatch로 LottoRank enum 값 찾아서 반환

MatchResult
- rank 멤버변수 추가
- getLottoRank를 활용하여 LottoRank 값 저장

WinningNumbers
- matchCount long -> int 변환

+++ src/main/java/lotto/domain/WinningNumbers.java
@@ -10,7 +10,7 @@ public class WinningNumbers {
     }
 
     public MatchResult match(Lotto lotto) {
-        long matchCount = lotto.countMatchNumbers(winningNumbers);
+        int matchCount = (int) lotto.countMatchNumbers(winningNumbers); // TODO long -> int
         boolean matchBonus = lotto.containsNumber(bonusNumber);
         return new MatchResult(matchCount, matchBonus);
     }
